split `notices_util.ts`
→ `util/my_date.ts`, `util/notices.ts`, `plugin/cli/index.ts`
